pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "agroops-copilot-api"
version = "0.1.0"
description = "FastAPI RAG copilot for agricultural operations with simulated field data."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "chromadb>=0.5.0",
    "httpx>=0.27.0",
    "pypdf>=4.2.0",
    "PyYAML>=6.0.1",
    "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0"
]
embeddings = [
    "sentence-transformers>=3.0.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

